refactor(daemon): route agent schema consumers through public boundaries
