+++ packages/daemon/src/agent-skills.ts
@@ -2,7 +2,7 @@ import { existsSync, lstatSync, readdirSync, realpathSync } from "node:fs";
 import { homedir } from "node:os";
 import path from "node:path";
 import { resolveHarnessLayout } from "../../kernel/src/index.ts";
-import type { AgentSkillDeclarationV1 } from "../../kernel/src/domain/agent-squad-schema.ts";
+import type { AgentSkillDeclarationV1 } from "../../kernel/src/index.ts";
 
 export interface ResolvedAgentSkill {
   readonly id: string;
